pyproject.toml
[tool.poetry]
name = "basic-api-users"
version = "0.1.0"
description = ""
authors = ["ronihdzz <ronaldo.runing_@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
loguru = "^0.7.3"
psycopg2-binary = "^2.9.10"
cryptography = "^44.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
coverage = "^7.6.0"
coverage-badge = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["./src"]